+++ dolphinscheduler-api/src/test/java/org/apache/dolphinscheduler/api/service/WorkerGroupServiceTest.java
@@ -17,6 +17,7 @@
 
 package org.apache.dolphinscheduler.api.service;
 
+import org.apache.dolphinscheduler.api.service.impl.WorkerGroupServiceImpl;
 import org.apache.dolphinscheduler.api.utils.PageInfo;
 import org.apache.dolphinscheduler.common.Constants;
 import org.apache.dolphinscheduler.common.enums.UserType;
@@ -49,7 +50,7 @@ public class WorkerGroupServiceTest {
     private static final Logger logger = LoggerFactory.getLogger(WorkerGroupServiceTest.class);
 
     @InjectMocks
-    private WorkerGroupService workerGroupService;
+    private WorkerGroupServiceImpl workerGroupService;
 
     @Mock
     private ProcessInstanceMapper processInstanceMapper;
